% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_distribution_age.R
\name{calcul_distribution_age}
\alias{calcul_distribution_age}
\title{Calcul de la distribution des âges des élus}
\usage{
calcul_distribution_age(df)
}
\arguments{
\item{df}{Un dataframe}
}
\value{
Un vecteur contenant les quantiles de l'âge.
}
\description{
La fonction calcule les quantiles 0, 25, 50, 75 et 100 de l'âge des élus
}
