% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compter_nombre_d_elus.R
\name{compter_nombre_d_elus}
\alias{compter_nombre_d_elus}
\title{Compter le nombre d'élus}
\usage{
compter_nombre_d_elus(df)
}
\arguments{
\item{df}{un dataframe}
}
\value{
le nombre d'élus
}
\description{
Elle compte le nombre unique d’élus basés sur le nom/prénom/date de naissance
}
