% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compter_nombres_adjoints.R
\name{compter_nombres_d_adjoints}
\alias{compter_nombres_d_adjoints}
\alias{compter_nombre_d_adjoints}
\title{Compter le nombres d'adjoints}
\usage{
compter_nombre_d_adjoints(df)
}
\arguments{
\item{df}{= un dataframe}
}
\value{
Le nombre d'adjoint unique
}
\description{
La fonction permet de donner le nombre d'adjoints unique en fonction de son nom, prénom et date de
naissance présente dans le dataframe
}
