% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_commune.R
\name{creer_commune}
\alias{creer_commune}
\title{Création d'un objet représentant une commune}
\usage{
creer_commune(df)
}
\arguments{
\item{df}{Un data.frame contenant les données des élus. Il doit inclure les colonnes suivantes :
\itemize{
\item \code{Code.de.la.commune} : Code de la commune
\item \code{Libellé.de.la.commune} : Nom de la commune
\item \code{Nom.de.l.élu} : Nom de l’élu
\item \code{Prénom.de.l.élu} : Prénom de l’élu
\item \code{Code.sexe} : Code du sexe de l’élu
\item \code{Date.de.naissance} : Date de naissance de l’élu
\item \code{Libellé.de.la.fonction} : Fonction de l’élu
\item \code{Date.de.début.de.la.fonction} : Date de début de la fonction
\item \code{Code.nationalité} : Code de la nationalité de l’élu
}}
}
\value{
Une liste de classe \code{"commune"} contenant les informations uniques sur les élus d'une commune.
}
\description{
Cette fonction crée une liste représentant une commune à partir des données des élus.
}
