% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_departement.R
\name{creer_departement}
\alias{creer_departement}
\title{Création d'un objet département}
\usage{
creer_departement(df)
}
\arguments{
\item{df}{Un \code{data.frame} contenant les données des communes d'un département. Il doit contenir au minimum les colonnes :
\itemize{
\item \code{Code.du.département} : code du département
\item \code{Libellé.du.département} : nom du département
\item \code{Code.de.la.commune} : code de la commune
\item \code{Libellé.de.la.commune} : nom de la commune
}}
}
\value{
Une liste de classe \code{"departement"} contenant :
\itemize{
\item \code{Code.du.département} : le code du département
\item \code{Libellé.du.département} : le nom du département
\item \code{Collectivite} : (si applicable) les informations sur la collectivité à statut particulier
\item \code{Communes} : une liste d'objets \code{commune}, un par commune du département
}
}
\description{
Cette fonction génère un objet de classe \code{departement} à partir d'un \code{data.frame} contenant des informations
sur les communes d'un département.
}
