% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_code_professions.R
\name{plot_code_professions}
\alias{plot_code_professions}
\title{plot du code de la profession}
\usage{
plot_code_professions(df)
}
\arguments{
\item{df}{Un dataframe}
}
\value{
un histogramme horizontal.
}
\description{
la fonction comptera, pour chaque code professionnel, le nombre d’élus ayant ce code. Un bar chart
horizontal est utiliser pour représenter ces informations.
}
