% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_commune.R
\name{plot_commune}
\alias{plot_commune}
\title{Visualisation des professions des élus d'une commune}
\usage{
plot_commune(df)
}
\arguments{
\item{df}{Un dataframe contenant les données des élus d'une commune. Il doit contenir les colonnes :
\itemize{
\item \code{Libellé.de.la.commune} : le nom de la commune
\item \code{Libellé.du.département} : le nom du département
\item \code{Code.de.la.catégorie.socio.professionnelle} : le code de la catégorie socio-professionnelle
}}
}
\value{
Un graphique à barres généré par \code{ggplot2} affichant les catégories socio-professionnelles des élus d'une commune.
}
\description{
Cette fonction crée un graphique à barres représentant la répartition des professions des élus d'une commune.
Le graphique affiche les codes professionnels des élus, leur fréquence et est filtré pour une seule commune.
}
