% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_departement.R
\name{plot_departement}
\alias{plot_departement}
\title{Visualisation des codes professionnels par département}
\usage{
plot_departement(df)
}
\arguments{
\item{df}{Data frame contenant les colonnes \code{Libellé.du.département}, \code{Libellé.de.la.commune}, et \code{Code.de.la.catégorie.socio.professionnelle}.}
}
\description{
Affiche un graphique des 10 codes professionnels les plus fréquents pour un département.
}
