% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_commune.R
\name{summary_commune}
\alias{summary_commune}
\title{Résumé de la commune}
\usage{
summary_commune(x)
}
\arguments{
\item{x}{un objet de type commune choisit pour avoir des informations}
}
\value{
Des inforformations sur la commune.
}
\description{
La fonction affiche les informations d’une commune donnée (nom, nombre d’élu·e·s, distribution des âges)
et identifie l’élu·e le/la plus âgé·e
}
