% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_departement.R
\name{summary_departement}
\alias{summary_departement}
\title{Résumé du département}
\usage{
summary_departement(x)
}
\arguments{
\item{x}{Un objet de type "département".}
}
\value{
Retourne les informations sur le département.
}
\description{
La fonction affiche les informations d’un département donné (nom, nombre de communes, nombre d’élu·e·s,
distribution des âges) et identifie l’élu·e le/la plus âgé·e et le/la plus jeune.
Elle calcule également la commune avec la moyenne d’âge la plus faible et la plus élevée.
}
