% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trouver_l_elu_le_plus_age.R
\name{trouver_l_elu_le_plus_age}
\alias{trouver_l_elu_le_plus_age}
\title{Trouver l'élu le plus agé}
\usage{
trouver_l_elu_le_plus_age(df)
}
\arguments{
\item{df}{= un dataframe}
}
\value{
l'age de l'élu le plus ancien ainsi que son nom et son prénom
}
\description{
La fonction permet de trouver l'elu le plus agé dans le dataframe
}
